% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_dados.R
\name{carregar_inputs}
\alias{carregar_inputs}
\title{carregar_inputs}
\usage{
carregar_inputs(arquivo_de_inputs = "params.xlsx",
  abas_a_ler = c("Variaveis"), nomes_inputs = c("Variaveis"))
}
\arguments{
\item{arquivo_de_inputs}{caminho para o arquivo de inputs com estratégias e incertezas}

\item{abas_a_ler}{abas a ler do arquivo de inputs}

\item{nomes_inputs}{Nome a ser atribuido aos dataframes de input.}
}
\value{
list com inputs para a simulação.
}
\description{
carregar_inputs
}
