% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_dados.R
\name{obter_dados_regionais_world_bank}
\alias{obter_dados_regionais_world_bank}
\title{obter_dados_regionais_world_bank
Esta função obtém dados do world bank utilizando a biblioteca wbstats. Os dados a serem coletados devem ser preenchidos na planilha de coleta de dados.}
\usage{
obter_dados_regionais_world_bank(arquivo_de_inputs)
}
\arguments{
\item{arquivo_de_inputs}{planilha de coleta de dados a ser utilizada}
}
\value{
list com dados coletados e uma tabela com regiões e seus códigos.
}
\description{
obter_dados_regionais_world_bank
Esta função obtém dados do world bank utilizando a biblioteca wbstats. Os dados a serem coletados devem ser preenchidos na planilha de coleta de dados.
}
