% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_dados.R
\name{obter_fundamentos_financeiros_quandl}
\alias{obter_fundamentos_financeiros_quandl}
\title{obter_fundamentos_financeiros_quandl
Função obtém dados de fundamentos financeiros usando a biblioteca quandl.}
\usage{
obter_fundamentos_financeiros_quandl(company_code = "DDD")
}
\arguments{
\item{company_code}{código da empresa constante na base de dados https://www.quandl.com/data/SF0-Free-US-Fundamentals-Data}
}
\value{
list com dados financeiros
}
\description{
obter_fundamentos_financeiros_quandl
Função obtém dados de fundamentos financeiros usando a biblioteca quandl.
}
